Update README.md
DOGE GPT is basically a BSC Token that is combination of AI and Cryptocurrency. The idea is to have use of DGPT in AI based platforms or applications that in future we may come up with. Our objective is to create a cutting-edge AI system that will assist investors in navigating the cryptocurrency market more efficiently and securely. We're dedicated to employ technology to address some of the trickiest issues facing crypto exchange platforms and investors.